Fixed: The UserAccount should be created if emailAddress is available
As mentioned in the IN parameter definition of create#PersonCustomer service, if newPassword field is empty, the UserAccount will be created and after this user will have to change or reset.

But with the exiting newPassword check, it was blocking the creation of UserAccount, thus removed the additional check of newPassword.